requirements.txt
discord.py>=2.4.0,<3
python-dotenv>=1.0.0,<2
httpx>=0.27.0,<1
beautifulsoup4>=4.12.0,<5
lxml>=5.0.0,<6
